# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/industries:
    get:
      tags:
        - Industries
      summary: List Industries
      description: Get a paginated list of Industries with optional filtering and sorting
      operationId: listIndustries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Industries
      summary: Create Industry
      description: Create a new Industry
      operationId: createIndustry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIndustryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Industry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/levels:
    get:
      tags:
        - Levels
      summary: List Levels
      description: Get a paginated list of Levels with optional filtering and sorting
      operationId: listLevels
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Levels
      summary: Create Level
      description: Create a new Level
      operationId: createLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLevelRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Level'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Industry:
      type: object
      required:
        - id
        - code
        - name
        - sector
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 5
        name:
          type: string
          maxLength: 300
        sector:
          type: string
        parent_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    CreateIndustryRequest:
      type: object
      required:
        - code
        - name
        - sector
        - status
        - metadata
      properties:
        code:
          type: string
          maxLength: 5
        name:
          type: string
          maxLength: 300
        sector:
          type: string
        parent_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    IndustryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Industry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Level:
      type: object
      required:
        - id
        - company_id
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        grade:
          nullable: true
          type: string
          maxLength: 20
        order_number:
          nullable: true
          type: integer
          format: int32
        metadata:
          type: object
...
    CreateLevelRequest:
      type: object
      required:
        - company_id
        - name
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        grade:
          nullable: true
          type: string
          maxLength: 20
        order_number:
          nullable: true
          type: integer
          format: int32
        metadata:
          type: object
...
    LevelList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Level'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []